backend/requirements.txt
annotated-types==0.7.0
anyio==4.9.0
certifi==2025.4.26
charset-normalizer==3.4.2
click==8.1.8
colorama==0.4.6
fastapi==0.115.12
h11==0.16.0
idna==3.10
networkx==3.1
pydantic==2.11.4
pydantic_core==2.33.2
python-dotenv==1.0.0
requests==2.32.3
sniffio==1.3.1
starlette==0.46.2
tweepy==4.14.0
typing-inspection==0.4.0
typing_extensions==4.13.2
urllib3==2.4.0
uvicorn==0.34.2
python-louvain==0.16
